+++ src/prelude.rs
@@ -8,4 +8,4 @@ pub use crate::error::{Error, Result};
 #[cfg(feature = "stream")]
 pub use crate::service::Streaming;
 pub use crate::service::{Files, Readiness, Tasks};
-pub use crate::{Client, ClientBuilder};
+pub use crate::{Client, ClientBuilder, Endpoint};